+++ pypugjs/ext/jinja.py
@@ -6,8 +6,8 @@ from jinja2.runtime import Undefined
 
 import pypugjs.runtime
 from pypugjs import Compiler as _Compiler
-from pypugjs.runtime import attrs as _attrs, iteration
-from pypugjs.utils import process,open
+from pypugjs.runtime import attrs as _attrs, iteration, open
+from pypugjs.utils import process
 
 ATTRS_FUNC = '__pypugjs_attrs'
 ITER_FUNC = '__pypugjs_iter'
