+++ pypugjs/utils.py
@@ -10,9 +10,8 @@ try:
 except Exception:
     izip, imap = zip, map
 from copy import deepcopy
-from chardet.universaldetector import UniversalDetector
 import six
-import io
+
 missing = object()
 
 
@@ -233,22 +232,3 @@ def process(src, filename=None, parser=Parser, compiler=HTMLCompiler, **kwargs):
     block = _parser.parse()
     _compiler = compiler(block, **kwargs)
     return _compiler.compile().strip()
-
-def open(file, mode='r', buffering=-1, encoding=None, errors=None,
-                 newline=None, closefd=True):
-        rawdata = io.open(file, mode='rb')
-
-        detector = UniversalDetector()
-        for line in rawdata.readlines():
-            detector.feed(line)
-            if detector.done: break
-        detector.close()
-        rawdata.close()
-
-        decoded = io.open(file, mode=mode, buffering=buffering,
-                          encoding=detector.result["encoding"],
-                          errors=errors,
-                          newline=newline,
-                          closefd=closefd)
-
-        return decoded
